Incorporate additional review comments
